lib.addresses: No exception if OUI list unavailable
Previously, we threw an exception if the OUI list text file was
unavailable.  This looks like it was a bug/regression from when
the OUI loading was redone and just hasn't been noticed for
nine years because the OUI list is generally loaded fine!  The
proper (now-restored) behavior is that it logs a message when
this happens and then carries on -- things will work, we just
won't be able to do OUI name resolution (which often doesn't
even matter).  The log message has also been slightly improved.