web: Don't allow logging random garbage
I think this is probably a problem with Python's web server logging
which we inherited, but now fix.

When there's a problem with an HTTP request, Python logs it.  But it
does this without seeming to do any validation or filtering at all.
This seems to mean you can get arbitrary 8-bit garbage sent to your
terminal, for example, which is not great.

While this could probably be used for nefarious reasons, it's also
just annoying, especially in the modern web era.  There are cases
where a browser may send HTTP/2, for example.  This looks like
nonsense to Python, so it logs it and ends up just spewing junk
into the log.

This commit does replacements for non-ASCII and control characters,
and also limits the length of what it tries to log.