web: CoreHandler redirects unknown URLs
When CoreHandler gets an unknown URL (to, say, path "/foo"), it now
redirects it with a trailing slash ("/foo/").  The purpose of this
is so that you can be sloppy and request, say, "/wslog", when you
*actually* mean "/wslog/".

This is part of a series of changes which includes adding trailing
slashes to existing handlers (e.g., the aforementioned "/wslog/").
See the following commits for the rationale and implications.