web: Major change to CookieGuard bounce page
CookieGuard has a "bounce" page.  If there's an attempt to fetch
a page without the guard cookie being set, the server sends back
the cookie, and redirects to the bounce page.  When the client
requests the bounce page, the cookie should be set, and the
bounce page redirects back to the original page (when things are
working correctly, anyway).

Two important and related questions are:
* What's the URL of the bounce page?
* How does the bounce back to the original page work (i.e.,
  how is the original URL carried forward)?

In the past, the bounce URL looked something like:
  /<_special_bounce_thing>?<original_url>

This is nice, straightforward, and clean.  However, it has at
least two shortcomings that both stem from the fact that the
prefix of the URL changes.

First, since the prefix of the URL has changed (we're now at
the "top level", not in any "subdirectory"), we may be in a
different authentication realm than the page you are trying to
access.  If you are doing anything more complex than trivial
site-wide authentication, the ramifications of this can vary
from slightly annoying, to functionality breaking, to
possibly creating a security vulnerability.

Second, if you are doing any sort of URL remapping (e.g., by
putting some POX web splits behind a reverse proxy server),
the bounce page may well not "just work".  Again, the
ramifications vary.  It could be slightly annoying (requiring
a bit of extra URL remapping) or quite difficult to deal with.

For these reasons, the bounce page is now invoked based on
a *suffix* of the URL, more like this:
  /<original_url>/<special_bounce_thing>

Whenever the POX web server sees a URL *ending* with the
special bounce string, it does a bounce page.  This is a sort
of odd thing, but it gets around the problems listed above,
and I think it works.  The unclean thing about it is that...
what if a page actually has a valid URL that happens to end
with the special bounce string?  This would obviously be bad,
but it seems very unlikely (the bounce string was chosen to
avoid some ways this could theoretically happen accidentally).
I also don't think it could be exploited to do anything *too*
nefarious (though I wouldn't swear to it).