web: Switch some 302 redirections to 307
There's some bad legacy with redirections.  It seems to have been a bug
that some old client implementations would change a POST to a GET when
they were redirected.  But that behavior got sort of codified.  Later
versions of HTTP have therefore introduced new redirections (307 and
308) which specifically maintain the original method.  307 (a temporary
redirection) was defined many years before 308 (permanent) for whatever
reason, so we'll use 307s just in case 308 isn't as widely implemented
as we'd like yet (it probably is, but it also probably doesn't matter
too much to us if we use temporary instead of permanent).

But the basic change here is that we switch to a new HTTP redirection
which has a better chance of properly redirecting POSTs.