web: Add trailing slashes to some URL handlers
This commit adds trailing slashes to existing handlers (e.g., instead
of handling "/wslog", we now handle "/wslog/".  It is part of a series
of commits, including one which redirects some URLs to versions with
trailing slashes, which means you may not really notice the change.

See the following commit(s) for the rationale and implications.