dnsd: Add ability to ignore some requests altogether
DNS amplifications attacks are still a thing in 2023.  While the
current server is not useful for them since it doesn't do
recursive resolutions, we don't really want to be part of that
whole thing anyway!  Thus, we can now take a list of names to
not only not resolve, but just not even respond to at all.
There's a built-in default list of some ones that are well
known / that I've observed.  Some of these are attacks and some
are probes looking for servers that can be *used* as attacks.