dnsd: Log source IP of DoH requests (with caveats)
Previously, log messages coming from DoH didn't have information
about the source IP.  They do now.  However, this is spoofable.
The reason is that we let the X-Forwarded-For header override
the actual remote IP address.  The reason is that this is a
common way for a proxy server to communicate the real remote IP
to the backend server.  But a malicious client could actually
set it itself, so you can't trust it too far.

In the future, we should perhaps allow X-Forwarded-For to
override the socket's remote IP if and only if the socket's
IP is a non-public IP (e.g., 10.0.0.0/8).