dnsd.web: Web UI/API work better with reverse proxies
The web UI tries to display the user's IP address to facilitate
creating records that point to themselves.  Similarly, the web API
tries to fill in the requester's IP address if one is not
specified for the same reason (and because this is what common
dynamic DNS services do).

Previously, it always did this by looking at the socket's remote
address.  However, this does not work if POX is running behind a
reverse proxy server -- it just gets the proxy's IP address.  A
common solution to this sort of thing is having the proxy fill in
the X-Forwarded-For header.  Thus, we now take the value from
there if it's set, and fall back to the socket's remote address
otherwise.  This is spoofable by the client, but there's no
point in doing so in these cases, since they could always have
just specified the IP address in the first place.

This is similar to another recent change to the DNS server where
it uses the same technique to get the IP address to use for
logging when handling DoH requests.  We should probably pull
this functionality out somewhere reusable, possibly with the
extra logic mentioned in the other commit.