proto.dhcp_client: Add PCap-based client
The DHCP client code was written as a reusable superclass, but we
only had a subclass for doing DHCP over OpenFlow.  This commit adds
a version which runs on an arbitrary real Ethernet-like interface
using pcap.  Actually setting up the interface hasn't been tested,
but it does properly get the lease.