proto.dhcp_client: Add reusable log messages
The DHCP client code works by raising various events.  This
commit adds a class which just has handlers for the events
which log them.  You could just add this as a superclass and
then override the methods you want to do more than log.
The pcap-based client now does this.