webcore/py3: Fix str/bytes with cgi.FieldStorage
I ran into this semi-recently while working on web stuff, but then
it seemed to disappear.  Today it was back.  The problem stems from
how FieldStorage handles binary and text, which matters in Python 3.
I actually have no idea how it's supposed to work, and I'm not sure
anyone else does either.  The class is slated for removal.  We'll
have to deal with that before too long, but for now, I've just got
it always opening the temp file in binary mode, which -- at least
for the case I was running into just now -- fixes things.  And it
seems like it should be okay, because the underlying socket is
always binary, but I'd definitely not swear to that.