web.jsonrpc: Try to conform more to specs
We now do a little manipulation of the result JSON to comply with
things the 1.0 and 2.0 spec say.  We also make sure there's always
a result value if one was expected (i.e., the 'id' was set in the
request) -- we return null if there wasn't one.

Finally, we now allow RPC handlers to not return anything at all.
This ends up getting treated as if they'd returned a dict with a
null result.