web: Replace deprecated cgi.FieldStorage
The Python cgi module has been deprecated since Python 3.11, and
will be removed in 3.13, which is due to be released shortly.  A
previous commit (a1873322) replaced cgi.escape() with html.escape(),
and another (9639fabe) replaced cgi.parse_header().

This commit replaces the last remaining usage of the cgi module in
webcore -- cgi.FieldStorage.  FieldStorage is replaced by a new class
that wraps Python's email.message.EmailMessage class to do the
parsing, and then provides enough of the FieldStorage interface to
fulfill webcore's needs.