web: Replace deprecated cgi.FieldStorage
The Python cgi module has been deprecated since Python 3.11, and
will be removed in 3.13, which is due to be released shortly.  A
previous commit (a1873322) replaced cgi.escape() with html.escape().
This commit replaces cgi.FieldStorage with a class that wraps the
email.message.EmailMessage class to do the parsing, and then
provides enough of the FieldStorage interface to fulfill POX's
needs.