web: InternalContentHandler can now handle JSON uploads
Previously, the POST handler for InternalContentHandler only
understood form data.  It now allows JSON too if the MIME type
of the upload is application/json.  The decoded JSON is passed
to the POST handler function.  You should be able to tell if
you got form data or JSON by inspecting the type of the 'data'
parameter.